Add naive C implementation of Damerau Levenshtein distance. Currently, this translates almost one to one jellyfish.damerau_levenshtein_distance, and is only marginally faster. damerau_levenshtein_google.c contains Google's implementation, which is slightly faster, but is currently unable to be compiled. Also, many package names were changed, as this is now a repository for multiple distances, instead of just hamming distance.

+++ c_distance/hamming.c
@@ -22,7 +22,7 @@ static PyObject* distance(PyObject *self, PyObject *args)
     int s1_len = strlen(s1);
     int s2_len = strlen(s2);
     long dist = s1_len - s2_len;
-    for (int i = 0; i < s2_len; i++)
+    for (size_t i = 0; i < s2_len; i++)
     {
         if (s1[i] != s2[i])
         {
